requirements.txt
# Gmail API
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Slack SDK
slack-sdk==3.26.1

# Environment and configuration
python-dotenv==1.0.0

# Scheduling
schedule==1.2.1

# Date/time utilities
python-dateutil==2.8.2
